requirements.txt
numpy==1.21.0
opencv-contrib-python==4.5.2.54
pyyaml==5.1
torch==1.8.0+cu101
